requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
black==20.8b1
click==7.1.2
Django==3.1.2
django-bootstrap4==2.3.1
django-dynamic-formsets==0.0.8
djangorestframework==3.12.2
geckodriver-autoinstaller==0.1.0
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==8.0.1
pytz==2020.1
regex==2020.10.28
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
